secure-ai-chatbot/backend/app/requirements.txt
# backend/app/requirements.txt
fastapi
python-dotenv
requests
uvicorn



